Schedulle Atributions
Initial table to select the Class to manage
Routes establishment

additional:
- intrusion logger function in teacher availabilitites;
- bug fix with empty ids array in bulk deletion/publishing

+++ Projecto/AgileWing/app/Http/Controllers/TeacherAvailabilityController.php
@@ -7,10 +7,11 @@ use App\TeacherAvailability;
 use App\User;
 use App\HourBlock;
 use App\AvailabilityType;
-use Illuminate\Support\Facades\Auth; //to get the logged in user
+use Illuminate\Support\Facades\Auth; //to get the loggedin user
 use Illuminate\Support\Facades\Validator; // to use custom rules in the validator
-use Carbon\Carbon;
-use App\Http\Controllers\Log;
+use Illuminate\Support\Facades\Log; // Import the Log facade
+use Carbon\Carbon; // to use the carbon methods (casts, addDay(), now(), etc)
+
 
 class TeacherAvailabilityController extends Controller
 {
@@ -115,6 +116,7 @@ class TeacherAvailabilityController extends Controller
      */
     public function store(Request $request)
     {
+        $this->detectIntrusion($request);
 
         $rules = [
             'user_id' => 'required|exists:users,id',
@@ -243,6 +245,7 @@ class TeacherAvailabilityController extends Controller
      */
     public function update(Request $request, $id)
     {
+        $this->detectIntrusion($request);
 
         // Check for published availability and user type permission to edit
         $teacherAvailability = TeacherAvailability::find($id);
@@ -295,6 +298,13 @@ class TeacherAvailabilityController extends Controller
 
     public function deleteSelected(Request $request)
     {
+        $this->detectIntrusion($request);
+
+        // Check for null or empty ids
+        $ids = $request->input('ids');
+        if (!$ids || !count($ids)) {
+            return redirect()->back()->with('error', 'Nenhum registo seleccionado para apagar.');
+        }
 
         try {
             TeacherAvailability::whereIn('id', $request->input('ids'))->delete();
@@ -306,6 +316,13 @@ class TeacherAvailabilityController extends Controller
 
     public function publishSelected(Request $request)
     {
+        $this->detectIntrusion($request);
+
+        // Check for null or empty ids
+        $ids = $request->input('ids');
+        if (!$ids || !count($ids)) {
+            return redirect()->back()->with('error', 'Nenhum registo seleccionado para publicar.');
+        }
 
         try {
             TeacherAvailability::whereIn('id', $request->input('ids'))->update(['is_locked' => true]);
@@ -314,26 +331,26 @@ class TeacherAvailabilityController extends Controller
             return redirect()->back()->with('error', 'Erro ao publicar os registos.');
         }
     }
-
+    
     /**
      * Auxiliary method to create () Handle the single date entry.
      */
     private function updateOrCreateSingleEntry($request, $date)
     {
-
+        
         $hourBlock = HourBlock::findOrFail($request->start_hour_block_id);
-
+        
         // Check for published availability and user type permission to edit
         $existingAvailability = TeacherAvailability::where([
             'user_id' => $request->user_id,
             'availability_date' => $date->toDateString(),
             'hour_block_id' => $hourBlock->id,
-        ])->first();
-
-        if ($existingAvailability && $existingAvailability->is_locked && auth()->user()->user_type_id !== 1) {
-            throw new \Exception('Alguns dos registos já se encontram publicados e não permitem edições.');
+            ])->first();
+            
+            if ($existingAvailability && $existingAvailability->is_locked && auth()->user()->user_type_id !== 1) {
+                throw new \Exception('Alguns dos registos já se encontram publicados e não permitem edições.');
         }
-
+        
         //UPDATE OR CREATE
         TeacherAvailability::updateOrCreate(
             [
@@ -343,10 +360,10 @@ class TeacherAvailabilityController extends Controller
             ],
             [
                 'availability_type_id' => $request->availability_type_id,
-            ]
-        );
-    }
-
+                ]
+            );
+        }
+        
     /**
      * Auxiliary method to create () Handle the date range entries.
      */
@@ -354,29 +371,29 @@ class TeacherAvailabilityController extends Controller
     {
         $startHourBlock = HourBlock::findOrFail($request->start_hour_block_id);
         $endHourBlock = HourBlock::findOrFail($request->end_hour_block_id);
-
+        
         // Gather all the hour block IDs
         $hourBlockIds = HourBlock::whereBetween('id', [$startHourBlock->id, $endHourBlock->id])
-                                 ->pluck('id')
-                                 ->toArray();
-
+            ->pluck('id')
+            ->toArray();
+            
         // Get all dates in the range.
         $dates = [];
         for ($date = $startDate->copy(); $date->lte($endDate); $date->addDay()) {
             $dates[] = $date->toDateString();
         }
-
+        
         // Check for any locked availabilities for the entire date and hour block range.
         $lockedRecords = TeacherAvailability::where('user_id', $request->user_id)
-                                           ->whereIn('availability_date', $dates)
-                                           ->whereIn('hour_block_id', $hourBlockIds)
-                                           ->where('is_locked', 1)
-                                           ->count();
-
+            ->whereIn('availability_date', $dates)
+            ->whereIn('hour_block_id', $hourBlockIds)
+            ->where('is_locked', 1)
+            ->count();
+        
         if ($lockedRecords > 0 && auth()->user()->user_type_id !== 1) {
             throw new \Exception('Alguns dos registos já se encontram publicados e não permitem edições.');
         }
-
+        
         // If there are no locked records, proceed to update or create the entries.
         foreach ($dates as $date) {
             foreach ($hourBlockIds as $hourBlockId) {
@@ -393,6 +410,33 @@ class TeacherAvailabilityController extends Controller
             }
         }
     }
-
-
+        
+    /**
+     * Detects unauthorized intrusion attempts based on user type and user ID.
+     *
+     * If an intrusion is detected:
+     * - The user's session is terminated.
+     * - An alert log is created with the IP address of the intruder.
+     * - A 403 Forbidden response is returned.
+     *
+     * @param  \Illuminate\Http\Request  $request  The incoming HTTP request instance.
+     * @throws \Symfony\Component\HttpKernel\Exception\HttpException  When an unauthorized action is detected.
+     * @return void
+     */
+    private function detectIntrusion($request)
+    {
+        if (auth()->user()->user_type_id !== 1 && auth()->user()->id != $request->user_id)
+        {
+            // End user's session
+            Auth::logout();
+            
+            // Log the intrusion
+            $ipAddress = $request->ip();
+            \Log::channel('intrusion')->alert("Intrusion Alert! Unauthorized access attempt from IP: $ipAddress");
+            
+            // Send 403 Forbidden response
+            abort(403, 'Unauthorized action.');
+        }
+    }   
 }
+
